Add Strapi authentication
adds a new file `strapi-api.ts` for implementing API calls to Strapi.
It encapsulate the auth implementation in its class and provide a few
example function for calling to strapi endpoint.

Bonus:
- refactored the auth to google-api, which also encapsulate the auth
functionality in the file and only expose the API functionality.
- refactored handler to use the new google-api and strapi-api.

Testing:
works locally with miniflare. additional testing is required for
cloudflare worker.

Signed-off-by: Daniel Huang <danielhuang37@outlook.com>